requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Agno Framework
agno==2.0.5

# AI/ML
google-generativeai==0.3.2
openai==1.6.1

# Data Processing
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2
PyPDF2==3.0.1
python-dateutil==2.8.2

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2

# File Processing
tabula-py==2.8.2
pdfplumber==0.10.3
xlrd==2.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
loguru==0.7.2